# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gestão de Encomendas API",
    "description": "API para gerenciamento de encomendas, vendas, despesas e produção artesanal.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/products/": {
      "post": {
        "summary": "Criar produto",
        "tags": [
          "Products"
        ],
        "description": "Cria um novo produto vinculado ao usuário autenticado.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string",
                    "minLength": 3
                  },
                  "descricao": {
                    "type": "string",
                    "minLength": 5
                  },
                  "preco_venda": {
                    "type": "number",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "preco_custo": {
                    "type": "number",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  },
                  "unidade_medida": {
                    "type": "string",
                    "minLength": 1
                  },
                  "quantidade_estoque": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "tempo_producao_minutos": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "imagem_url": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "uri"
                      },
                      {
                        "type": "string",
                        "enum": [
                          ""
                        ]
                      }
                    ]
                  },
                  "categoria": {
                    "type": "string",
                    "minLength": 3
                  }
                },
                "required": [
                  "nome",
                  "descricao",
                  "preco_venda",
                  "categoria"
                ]
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "user_id": {
                      "type": "number"
                    },
                    "nome": {
                      "type": "string"
                    },
                    "descricao": {
                      "type": "string"
                    },
                    "preco_venda": {
                      "type": "number"
                    },
                    "preco_custo": {
                      "nullable": true,
                      "type": "number"
                    },
                    "unidade_medida": {
                      "nullable": true,
                      "type": "string"
                    },
                    "quantidade_estoque": {
                      "nullable": true,
                      "type": "number"
                    },
                    "tempo_producao_minutos": {
                      "nullable": true,
                      "type": "number"
                    },
                    "imagem_url": {
                      "nullable": true,
                      "type": "string"
                    },
                    "categoria": {
                      "type": "string"
                    },
                    "ativo": {
                      "type": "boolean"
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "updated_at": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "user_id",
                    "nome",
                    "descricao",
                    "preco_venda",
                    "categoria"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Erro de validação"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado — token inválido, expirado ou usuário sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Não autorizado — token inválido, expirado ou usuário sem permissão"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Recurso não encontrado"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Erro interno do servidor"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Listar produtos do usuário",
        "tags": [
          "Products"
        ],
        "description": "Retorna todos os produtos do usuário autenticado.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "user_id": {
                        "type": "number"
                      },
                      "nome": {
                        "type": "string"
                      },
                      "descricao": {
                        "type": "string"
                      },
                      "preco_venda": {
                        "type": "number"
                      },
                      "preco_custo": {
                        "nullable": true,
                        "type": "number"
                      },
                      "unidade_medida": {
                        "nullable": true,
                        "type": "string"
                      },
                      "quantidade_estoque": {
                        "nullable": true,
                        "type": "number"
                      },
                      "tempo_producao_minutos": {
                        "nullable": true,
                        "type": "number"
                      },
                      "imagem_url": {
                        "nullable": true,
                        "type": "string"
                      },
                      "categoria": {
                        "type": "string"
                      },
                      "ativo": {
                        "type": "boolean"
                      },
                      "created_at": {
                        "type": "string"
                      },
                      "updated_at": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "user_id",
                      "nome",
                      "descricao",
                      "preco_venda",
                      "categoria"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Erro de validação"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado — token inválido, expirado ou usuário sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Não autorizado — token inválido, expirado ou usuário sem permissão"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Recurso não encontrado"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Erro interno do servidor"
                }
              }
            }
          }
        }
      }
    },
...
    "/clients/": {
      "post": {
        "summary": "Criar cliente",
        "tags": [
          "Clients"
        ],
        "description": "Cadastra um novo cliente vinculado ao usuário autenticado.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string",
                    "minLength": 3
                  },
                  "telefone": {
                    "type": "string",
                    "pattern": "^\\(?\\d{2}\\)?\\s?\\d{4,5}-?\\d{4}$"
                  },
                  "email": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "email",
                        "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                      },
                      {
                        "type": "string",
                        "enum": [
                          ""
                        ]
                      }
                    ]
                  },
                  "endereco": {
                    "type": "string",
                    "minLength": 5
                  }
                },
                "required": [
                  "nome",
                  "telefone",
                  "endereco"
                ]
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "user_id": {
                      "type": "number"
                    },
                    "nome": {
                      "type": "string"
                    },
                    "telefone": {
                      "type": "string"
                    },
                    "email": {
                      "nullable": true,
                      "type": "string"
                    },
                    "endereco": {
                      "type": "string"
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "updated_at": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "user_id",
                    "nome",
                    "telefone",
                    "endereco"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Erro de validação"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado — token inválido, expirado ou usuário sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Não autorizado — token inválido, expirado ou usuário sem permissão"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Recurso não encontrado"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Erro interno do servidor"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Listar clientes do usuário",
        "tags": [
          "Clients"
        ],
        "description": "Retorna todos os clientes do usuário autenticado.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number"
                      },
                      "user_id": {
                        "type": "number"
                      },
                      "nome": {
                        "type": "string"
                      },
                      "telefone": {
                        "type": "string"
                      },
                      "email": {
                        "nullable": true,
                        "type": "string"
                      },
                      "endereco": {
                        "type": "string"
                      },
                      "created_at": {
                        "type": "string"
                      },
                      "updated_at": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "user_id",
                      "nome",
                      "telefone",
                      "endereco"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Erro de validação"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado — token inválido, expirado ou usuário sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Não autorizado — token inválido, expirado ou usuário sem permissão"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Recurso não encontrado"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "description": "Erro interno do servidor"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    }
  ],
...
}